Renamed and added AddApplicationServices method
Renamed to application service collection extensions, added a add application services method
